requirements.txt
huggingface_hub==0.33.0
matplotlib==3.10.0
numpy==2.0.2
pandas==2.2.2
pillow==11.2.1

# CPU-only PyTorch builds
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu

# For loading local models
timm==1.0.15

tqdm==4.67.1
transformers==4.52.4
easyocr==1.7.2

pymupdf

# API dependencies
flask
flask-cors

# Python notebooks (optional, for development)
ipykernel